ci: paramaterize async/sync clients and remove duplicate async test suite
