+++ .github/workflows/integration_test.yml
@@ -17,6 +17,7 @@ jobs:
     strategy:
       matrix:
         python-version: ["3.9", "3.10", "3.11", "3.12"]
+        async: ["--sync", "--async"]
 
         
     steps:
@@ -60,7 +61,7 @@ jobs:
 
       working-directory: tests
 
-    - name: Run unit tests
+    - name: Run integration tests
       run: |
         docker run -d  --network=host -p 5000:5000 --name aerospike-vector-search -v ./aerospike-vector-search.yml:/etc/aerospike-vector-search/aerospike-vector-search.yml -v ./features.conf:/etc/aerospike-vector-search/features.conf aerospike/aerospike-vector-search:1.0.0
     
@@ -70,7 +71,7 @@ jobs:
         sleep 5
 
         docker ps
-        python -m pytest standard -s --host 0.0.0.0 --port 5000 --cov=aerospike_vector_search
+        python -m pytest standard -s --host 0.0.0.0 --port 5000 --cov=aerospike_vector_search ${{ matrix.async }}
 
         mv .coverage coverage_data
       working-directory: tests
