--- tests/standard/aio/requirements.txt
@@ -1,4 +1,0 @@
-numpy==1.26.4
-pytest==7.4.0
-pytest-aio==1.5.0
-..
