--- tests/standard/aio/test_vector_client_exists.py
@@ -1,68 +1,0 @@
-import pytest
-import grpc
-
-from aerospike_vector_search import AVSServerError
-from utils import DEFAULT_NAMESPACE
-
-from hypothesis import given, settings, Verbosity
-
-
-class exists_test_case:
-    def __init__(
-        self,
-        *,
-        namespace,
-        set_name,
-        timeout,
-    ):
-        self.namespace = namespace
-        self.set_name = set_name
-        self.timeout = timeout
-
-
-#@given(random_key=key_strategy())
-#@settings(max_examples=1, deadline=1000)
-@pytest.mark.parametrize(
-    "test_case",
-    [
-        exists_test_case(
-            namespace=DEFAULT_NAMESPACE,
-            set_name=None,
-            timeout=None,
-        ),
-        exists_test_case(
-            namespace=DEFAULT_NAMESPACE,
-            set_name=None,
-            timeout=None,
-        ),
-    ],
-)
-async def test_vector_exists(session_vector_client, test_case, record):
-    result = await session_vector_client.exists(
-        namespace=test_case.namespace,
-        key=record,
-    )
-    assert result is True
-
-
-#@given(random_key=key_strategy())
-#@settings(max_examples=1, deadline=1000)
-@pytest.mark.parametrize(
-    "test_case",
-    [
-        exists_test_case(
-            namespace="test", set_name=None, timeout=0.0001
-        ),
-    ],
-)
-async def test_vector_exists_timeout(
-    session_vector_client, test_case, record, with_latency
-):
-    if not with_latency:
-        pytest.skip("Server latency too low to test timeout")
-    with pytest.raises(AVSServerError) as e_info:
-        for i in range(10):
-            result = await session_vector_client.exists(
-                namespace=test_case.namespace, key=record, timeout=test_case.timeout
-            )
-    assert e_info.value.rpc_error.code() == grpc.StatusCode.DEADLINE_EXCEEDED
