+++ tests/standard/test_admin_client_index_create.py
@@ -1,11 +1,10 @@
 import pytest
-from aerospike_vector_search import types, AVSServerError
 import grpc
 
-from ...utils import random_name, DEFAULT_NAMESPACE
+from aerospike_vector_search import types, AVSServerError
+from utils import random_name, drop_specified_index, DEFAULT_NAMESPACE
 
-from .aio_utils import drop_specified_index
-from hypothesis import given, settings, Verbosity, Phase
+from hypothesis import given, settings, Verbosity
 
 server_defaults = {
     "m": 16,
@@ -46,7 +45,7 @@ class index_create_test_case:
 
 
 #@given(random_name=index_strategy())
-#@settings(max_examples=1, deadline=1000, phases=(Phase.generate,))
+#@settings(max_examples=1, deadline=1000)
 @pytest.mark.parametrize(
     "test_case",
     [
@@ -60,13 +59,17 @@ class index_create_test_case:
             index_labels=None,
             index_storage=None,
             timeout=None,
-        ),
+        )
     ],
 )
-async def test_index_create(session_admin_client, test_case, random_name):
-    if test_case == None:
-        return
-    await session_admin_client.index_create(
+def test_index_create(session_admin_client, test_case, random_name):
+    try:
+        session_admin_client.index_drop(namespace=DEFAULT_NAMESPACE, name=random_name)
+    except AVSServerError as se:
+        if se.rpc_error.code() != grpc.StatusCode.NOT_FOUND:
+            pass
+
+    session_admin_client.index_create(
         namespace=test_case.namespace,
         name=random_name,
         vector_field=test_case.vector_field,
@@ -78,7 +81,8 @@ async def test_index_create(session_admin_client, test_case, random_name):
         index_storage=test_case.index_storage,
         timeout=test_case.timeout,
     )
-    results = await session_admin_client.index_list()
+
+    results = session_admin_client.index_list()
     found = False
     for result in results:
         if result["id"]["name"] == random_name:
@@ -91,16 +95,16 @@ async def test_index_create(session_admin_client, test_case, random_name):
             assert result["hnsw_params"]["ef"] == 100
             assert result["hnsw_params"]["batching_params"]["max_index_records"] == 100000
             assert result["hnsw_params"]["batching_params"]["index_interval"] == 30000
-            assert result["hnsw_params"]["batching_params"]["max_reindex_records"] == max(100000 / 10, 1000)
+            assert result["hnsw_params"]["batching_params"]["max_reindex_records"] == max(100000/10, 1000)
             assert result["hnsw_params"]["batching_params"]["reindex_interval"] == 30000
             assert result["storage"]["namespace"] == test_case.namespace
             assert result["storage"]["set_name"] == random_name
     assert found == True
-    await drop_specified_index(session_admin_client, test_case.namespace, random_name)
+    drop_specified_index(session_admin_client, test_case.namespace, random_name)
 
 
 #@given(random_name=index_strategy())
-#@settings(max_examples=1, deadline=1000, phases=(Phase.generate,))
+#@settings(max_examples=1, deadline=1000)
 @pytest.mark.parametrize(
     "test_case",
     [
@@ -128,10 +132,15 @@ async def test_index_create(session_admin_client, test_case, random_name):
         ),
     ],
 )
-async def test_index_create_with_dimnesions(
-    session_admin_client, test_case, random_name
-):
-    await session_admin_client.index_create(
+def test_index_create_with_dimnesions(session_admin_client, test_case, random_name):
+
+    try:
+        session_admin_client.index_drop(namespace=DEFAULT_NAMESPACE, name=random_name)
+    except AVSServerError as se:
+        if se.rpc_error.code() != grpc.StatusCode.NOT_FOUND:
+            pass
+
+    session_admin_client.index_create(
         namespace=test_case.namespace,
         name=random_name,
         vector_field=test_case.vector_field,
@@ -144,7 +153,7 @@ async def test_index_create_with_dimnesions(
         timeout=test_case.timeout,
     )
 
-    results = await session_admin_client.index_list()
+    results = session_admin_client.index_list()
 
     found = False
     for result in results:
@@ -165,11 +174,11 @@ async def test_index_create_with_dimnesions(
             assert result["storage"]["set_name"] == random_name
     assert found == True
 
-    await drop_specified_index(session_admin_client, test_case.namespace, random_name)
+    drop_specified_index(session_admin_client, test_case.namespace, random_name)
 
 
 #@given(random_name=index_strategy())
-#@settings(max_examples=1, deadline=1000, phases=(Phase.generate,))
+#@settings(max_examples=1, deadline=1000)
 @pytest.mark.parametrize(
     "test_case",
     [
@@ -219,11 +228,17 @@ async def test_index_create_with_dimnesions(
         ),
     ],
 )
-async def test_index_create_with_vector_distance_metric(
+def test_index_create_with_vector_distance_metric(
     session_admin_client, test_case, random_name
 ):
 
-    await session_admin_client.index_create(
+    try:
+        session_admin_client.index_drop(namespace=DEFAULT_NAMESPACE, name=random_name)
+    except AVSServerError as se:
+        if se.rpc_error.code() != grpc.StatusCode.NOT_FOUND:
+            pass
+
+    session_admin_client.index_create(
         namespace=test_case.namespace,
         name=random_name,
         vector_field=test_case.vector_field,
@@ -235,7 +250,7 @@ async def test_index_create_with_vector_distance_metric(
         index_storage=test_case.index_storage,
         timeout=test_case.timeout,
     )
-    results = await session_admin_client.index_list()
+    results = session_admin_client.index_list()
     found = False
     for result in results:
         if result["id"]["name"] == random_name:
@@ -253,11 +268,11 @@ async def test_index_create_with_vector_distance_metric(
             assert result["storage"]["namespace"] == test_case.namespace
             assert result["storage"]["set_name"] == random_name
     assert found == True
-    await drop_specified_index(session_admin_client, test_case.namespace, random_name)
+    drop_specified_index(session_admin_client, test_case.namespace, random_name)
 
 
 #@given(random_name=index_strategy())
-#@settings(max_examples=1, deadline=1000, phases=(Phase.generate,))
+#@settings(max_examples=1, deadline=1000)
 @pytest.mark.parametrize(
     "test_case",
     [
@@ -285,9 +300,15 @@ async def test_index_create_with_vector_distance_metric(
         ),
     ],
 )
-async def test_index_create_with_sets(session_admin_client, test_case, random_name):
+def test_index_create_with_sets(session_admin_client, test_case, random_name):
 
-    await session_admin_client.index_create(
+    try:
+        session_admin_client.index_drop(namespace=DEFAULT_NAMESPACE, name=random_name)
+    except AVSServerError as se:
+        if se.rpc_error.code() != grpc.StatusCode.NOT_FOUND:
+            pass
+
+    session_admin_client.index_create(
         namespace=test_case.namespace,
         name=random_name,
         vector_field=test_case.vector_field,
@@ -299,7 +320,7 @@ async def test_index_create_with_sets(session_admin_client, test_case, random_na
         index_storage=test_case.index_storage,
         timeout=test_case.timeout,
     )
-    results = await session_admin_client.index_list()
+    results = session_admin_client.index_list()
     found = False
     for result in results:
         if result["id"]["name"] == random_name:
@@ -317,11 +338,11 @@ async def test_index_create_with_sets(session_admin_client, test_case, random_na
             assert result["storage"]["namespace"] == test_case.namespace
             assert result["storage"]["set_name"] == random_name
     assert found == True
-    await drop_specified_index(session_admin_client, test_case.namespace, random_name)
+    drop_specified_index(session_admin_client, test_case.namespace, random_name)
 
 
 #@given(random_name=index_strategy())
-#@settings(max_examples=1, deadline=1000, phases=(Phase.generate,))
+#@settings(max_examples=1, deadline=1000)
 @pytest.mark.parametrize(
     "test_case",
     [
@@ -330,12 +351,12 @@ async def test_index_create_with_sets(session_admin_client, test_case, random_na
             vector_field="example_10",
             dimensions=1024,
             vector_distance_metric=None,
-            sets=None,
+            sets="demo",
             index_params=types.HnswParams(
                 m=32,
                 ef_construction=200,
                 ef=400,
-                enable_vector_integrity_check= False,
+                enable_vector_integrity_check = False,
             ),
             index_labels=None,
             index_storage=None,
@@ -346,12 +367,22 @@ async def test_index_create_with_sets(session_admin_client, test_case, random_na
             vector_field="example_11",
             dimensions=1024,
             vector_distance_metric=None,
+            sets="demo",
+            index_params=types.HnswParams(
+                m=8, ef_construction=50, ef=25, max_mem_queue_size=16384, enable_vector_integrity_check = True,
+            ),
+            index_labels=None,
+            index_storage=None,
+            timeout=None,
+        ),
+        index_create_test_case(
+            namespace=DEFAULT_NAMESPACE,
+            vector_field="example_20",
+            dimensions=1024,
+            vector_distance_metric=None,
             sets=None,
             index_params=types.HnswParams(
-                m=8,
-                ef_construction=50,
-                ef=25,
-                enable_vector_integrity_check= True
+                m=8, enable_vector_integrity_check= True,
             ),
             index_labels=None,
             index_storage=None,
@@ -362,21 +393,7 @@ async def test_index_create_with_sets(session_admin_client, test_case, random_na
             vector_field="example_12",
             dimensions=1024,
             vector_distance_metric=None,
-            sets=None,
-            index_params=types.HnswParams(
-                m=8,
-                enable_vector_integrity_check= True,
-            ),
-            index_labels=None,
-            index_storage=None,
-            timeout=None,
-        ),
-        index_create_test_case(
-            namespace=DEFAULT_NAMESPACE,
-            vector_field="example_13",
-            dimensions=1024,
-            vector_distance_metric=None,
-            sets=None,
+            sets="demo",
             index_params=types.HnswParams(
                 batching_params=types.HnswBatchingParams(max_index_records=2000, index_interval=20000, max_reindex_records=1500, reindex_interval=70000),
                 enable_vector_integrity_check= True
@@ -387,7 +404,7 @@ async def test_index_create_with_sets(session_admin_client, test_case, random_na
         ),
         index_create_test_case(
             namespace=DEFAULT_NAMESPACE,
-            vector_field="example_20",
+            vector_field="example_13",
             dimensions=1024,
             vector_distance_metric=None,
             sets="demo",
@@ -412,10 +429,13 @@ async def test_index_create_with_sets(session_admin_client, test_case, random_na
         ),
     ],
 )
-async def test_index_create_with_index_params(
-    session_admin_client, test_case, random_name
-):
-    await session_admin_client.index_create(
+def test_index_create_with_index_params(session_admin_client, test_case, random_name):
+    try:
+        session_admin_client.index_drop(namespace=DEFAULT_NAMESPACE, name=random_name)
+    except AVSServerError as se:
+        if se.rpc_error.code() != grpc.StatusCode.NOT_FOUND:
+            pass
+    session_admin_client.index_create(
         namespace=test_case.namespace,
         name=random_name,
         vector_field=test_case.vector_field,
@@ -427,7 +447,8 @@ async def test_index_create_with_index_params(
         index_storage=test_case.index_storage,
         timeout=test_case.timeout,
     )
-    results = await session_admin_client.index_list()
+
+    results = session_admin_client.index_list()
     found = False
     for result in results:
         if result["id"]["name"] == random_name:
@@ -438,11 +459,15 @@ async def test_index_create_with_index_params(
             assert result["hnsw_params"]["m"] == test_case.index_params.m or server_defaults
             assert (
                 result["hnsw_params"]["ef_construction"]
-                == test_case.index_params.ef_construction or server_defaults
+                == test_case.index_params.ef_construction  or server_defaults
             )
             assert result["hnsw_params"]["ef"] == test_case.index_params.ef or server_defaults
-            assert result["hnsw_params"][
-                       "enable_vector_integrity_check"] == test_case.index_params.enable_vector_integrity_check
+            if getattr(result.hnsw_params, 'max_mem_queue_size', None) is not None:
+                assert (
+                    result["hnsw_params"]["max_mem_queue_size"]
+                    == test_case.index_params.max_mem_queue_size or server_defaults
+                )
+
             assert (
                     result["hnsw_params"]["batching_params"]["max_index_records"]
                     == test_case.index_params.batching_params.max_index_records or server_defaults
@@ -452,29 +477,74 @@ async def test_index_create_with_index_params(
                     == test_case.index_params.batching_params.index_interval or server_defaults
             )
             assert (
-                   result["hnsw_params"]["batching_params"]["max_reindex_records"]
-                   == test_case.index_params.batching_params.max_reindex_records or server_defaults
+                    result["hnsw_params"]["batching_params"]["max_reindex_records"]
+                    == test_case.index_params.batching_params.max_reindex_records or server_defaults
             )
-
             assert (
-                   result["hnsw_params"]["batching_params"]["reindex_interval"]
-                   == test_case.index_params.batching_params.reindex_interval or server_defaults
+                    result["hnsw_params"]["batching_params"]["reindex_interval"]
+                    == test_case.index_params.batching_params.reindex_interval or server_defaults
             )
 
+            assert result["hnsw_params"][
+                       "enable_vector_integrity_check"] == test_case.index_params.enable_vector_integrity_check or server_defaults
+            """
+            if getattr(result.hnsw_params, 'index_caching_params', None) is not None:
+
+                assert (
+                    int(result["hnsw_params"]["index_caching_params"]["max_entries"])
+                    == test_case.index_params.index_caching_params.max_entries
+                )
+                assert (
+                    int(result["hnsw_params"]["index_caching_params"]["expiry"])
+                    == test_case.index_params.index_caching_params.expiry
+                )
+            if getattr(result.hnsw_params, 'merge_params', None) is not None:
+                assert (
+                    result["hnsw_params"]["merge_params"]["index_parallelism"]
+                    == test_case.index_params.merge_params.index_parallelism
+                )
+                assert (
+                    result["hnsw_params"]["merge_params"]["reindex_parallelism"]
+                    == test_case.index_params.merge_params.reindex_parallelism
+                )
+            if getattr(result.hnsw_params, 'healer_params', None) is not None:
+                assert (
+                    int(
+                        result["hnsw_params"]["healer_params"]["max_scan_rate_per_node"]
+                    )
+                    == test_case.index_params.healer_params.max_scan_rate_per_node
+                )
+                assert (
+                    int(result["hnsw_params"]["healer_params"]["max_scan_page_size"])
+                    == test_case.index_params.healer_params.max_scan_page_size
+                )
+                assert (
+                    int(result["hnsw_params"]["healer_params"]["re_index_percent"])
+                    == test_case.index_params.healer_params.re_index_percent
+                )
+                assert (
+                    result["hnsw_params"]["healer_params"]["schedule"]
+                    == test_case.index_params.healer_params.schedule
+                )
+                assert (
+                    result["hnsw_params"]["healer_params"]["parallelism"]
+                    == test_case.index_params.healer_params.parallelism
+                )
+            """
             assert result["storage"]["namespace"] == test_case.namespace
             assert result["storage"]["set_name"] == random_name
     assert found == True
-    await drop_specified_index(session_admin_client, test_case.namespace, random_name)
+    drop_specified_index(session_admin_client, test_case.namespace, random_name)
 
 
 #@given(random_name=index_strategy())
-#@settings(max_examples=1, deadline=1000, phases=(Phase.generate,))
+#@settings(max_examples=1, deadline=1000)
 @pytest.mark.parametrize(
     "test_case",
     [
         index_create_test_case(
             namespace=DEFAULT_NAMESPACE,
-            vector_field="example_14",
+            vector_field="example_16",
             dimensions=1024,
             vector_distance_metric=None,
             sets=None,
@@ -482,13 +552,16 @@ async def test_index_create_with_index_params(
             index_labels={"size": "large", "price": "$4.99", "currencyType": "CAN"},
             index_storage=None,
             timeout=None,
-        ),
+        )
     ],
 )
-async def test_index_create_index_labels(session_admin_client, test_case, random_name):
-    if test_case == None:
-        return
-    await session_admin_client.index_create(
+def test_index_create_index_labels(session_admin_client, test_case, random_name):
+    try:
+        session_admin_client.index_drop(namespace=DEFAULT_NAMESPACE, name=random_name)
+    except AVSServerError as se:
+        if se.rpc_error.code() != grpc.StatusCode.NOT_FOUND:
+            pass
+    session_admin_client.index_create(
         namespace=test_case.namespace,
         name=random_name,
         vector_field=test_case.vector_field,
@@ -500,7 +573,8 @@ async def test_index_create_index_labels(session_admin_client, test_case, random
         index_storage=test_case.index_storage,
         timeout=test_case.timeout,
     )
-    results = await session_admin_client.index_list()
+
+    results = session_admin_client.index_list()
     found = False
     for result in results:
         if result["id"]["name"] == random_name:
@@ -508,6 +582,8 @@ async def test_index_create_index_labels(session_admin_client, test_case, random
             assert result["id"]["namespace"] == test_case.namespace
             assert result["dimensions"] == test_case.dimensions
             assert result["field"] == test_case.vector_field
+            assert isinstance(result["field"], str)
+
             assert result["hnsw_params"]["m"] == 16
             assert result["hnsw_params"]["ef_construction"] == 100
             assert result["hnsw_params"]["ef"] == 100
@@ -515,20 +591,21 @@ async def test_index_create_index_labels(session_admin_client, test_case, random
             assert result["hnsw_params"]["batching_params"]["index_interval"] == 30000
             assert result["hnsw_params"]["batching_params"]["max_reindex_records"] == max(100000 / 10, 1000)
             assert result["hnsw_params"]["batching_params"]["reindex_interval"] == 30000
+
             assert result["storage"]["namespace"] == test_case.namespace
             assert result["storage"]["set_name"] == random_name
     assert found == True
-    await drop_specified_index(session_admin_client, test_case.namespace, random_name)
+    drop_specified_index(session_admin_client, test_case.namespace, random_name)
 
 
 #@given(random_name=index_strategy())
-#@settings(max_examples=1, deadline=1000, phases=(Phase.generate,))
+#@settings(max_examples=1, deadline=1000)
 @pytest.mark.parametrize(
     "test_case",
     [
         index_create_test_case(
             namespace=DEFAULT_NAMESPACE,
-            vector_field="example_15",
+            vector_field="example_17",
             dimensions=1024,
             vector_distance_metric=None,
             sets=None,
@@ -539,8 +616,13 @@ async def test_index_create_index_labels(session_admin_client, test_case, random
         ),
     ],
 )
-async def test_index_create_index_storage(session_admin_client, test_case, random_name):
-    await session_admin_client.index_create(
+def test_index_create_index_storage(session_admin_client, test_case, random_name):
+    try:
+        session_admin_client.index_drop(namespace=DEFAULT_NAMESPACE, name=random_name)
+    except AVSServerError as se:
+        if se.rpc_error.code() != grpc.StatusCode.NOT_FOUND:
+            pass
+    session_admin_client.index_create(
         namespace=test_case.namespace,
         name=random_name,
         vector_field=test_case.vector_field,
@@ -552,14 +634,15 @@ async def test_index_create_index_storage(session_admin_client, test_case, rando
         index_storage=test_case.index_storage,
         timeout=test_case.timeout,
     )
-    results = await session_admin_client.index_list()
+
+    results = session_admin_client.index_list()
     found = False
     for result in results:
         if result["id"]["name"] == random_name:
             found = True
             assert result["id"]["namespace"] == test_case.namespace
             assert result["dimensions"] == test_case.dimensions
-            assert result["field"] == test_case.vector_field
+            assert isinstance(result["field"], str)
             assert result["hnsw_params"]["m"] == 16
             assert result["hnsw_params"]["ef_construction"] == 100
             assert result["hnsw_params"]["ef"] == 100
@@ -573,13 +656,13 @@ async def test_index_create_index_storage(session_admin_client, test_case, rando
 
 
 #@given(random_name=index_strategy())
-#@settings(max_examples=1, deadline=1000, phases=(Phase.generate,))
+#@settings(max_examples=1, deadline=1000)
 @pytest.mark.parametrize(
     "test_case",
     [
         index_create_test_case(
             namespace=DEFAULT_NAMESPACE,
-            vector_field="example_16",
+            vector_field="example_18",
             dimensions=1024,
             vector_distance_metric=None,
             sets=None,
@@ -590,17 +673,21 @@ async def test_index_create_index_storage(session_admin_client, test_case, rando
         ),
     ],
 )
-async def test_index_create_timeout(
+def test_index_create_timeout(
     session_admin_client, test_case, random_name, with_latency
 ):
 
     if not with_latency:
         pytest.skip("Server latency too low to test timeout")
+    try:
+        session_admin_client.index_drop(namespace=DEFAULT_NAMESPACE, name=random_name)
+    except AVSServerError as se:
+        if se.rpc_error.code() != grpc.StatusCode.NOT_FOUND:
+            pass
 
-    with pytest.raises(AVSServerError) as e_info:
-        for i in range(10):
-
-            await session_admin_client.index_create(
+    for i in range(10):
+        try:
+            session_admin_client.index_create(
                 namespace=test_case.namespace,
                 name=random_name,
                 vector_field=test_case.vector_field,
@@ -612,4 +699,8 @@ async def test_index_create_timeout(
                 index_storage=test_case.index_storage,
                 timeout=test_case.timeout,
             )
-    assert e_info.value.rpc_error.code() == grpc.StatusCode.DEADLINE_EXCEEDED
+        except AVSServerError as se:
+            if se.rpc_error.code() == grpc.StatusCode.DEADLINE_EXCEEDED:
+                assert se.rpc_error.code() == grpc.StatusCode.DEADLINE_EXCEEDED
+                return
+    assert "In several attempts, the timeout did not happen" == "TEST FAIL"
