+++ tests/standard/test_admin_client_index_drop.py
