+++ tests/standard/test_vector_client_delete.py
