+++ tests/standard/test_vector_client_exists.py
