+++ tests/standard/test_vector_client_get.py
