+++ tests/standard/test_vector_client_insert.py
